requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
google-cloud-firestore>=2.11.0
google-auth>=2.23.0

